// webapp/src/api/model/weapon.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Dice } from './dice';
import type { WeaponEffectsItem } from './weaponEffectsItem';
import type { Material } from './material';
import type { ERarity } from './eRarity';

export interface Weapon {
  damage: number;
  description: string;
  dice: Dice;
  effects: WeaponEffectsItem[];
  hit: number;
  id?: string;
  initiative: number;
  material: Material;
  maximumStackSize: number;
  minimumStackSize: number;
  name: string;
  note: string;
  rarity: ERarity;
  requirement: string;
  tags: string[];
  tier: number;
  upgradeSlots: number;
  vendorPrice: number;
}